Move the Core\View Facade to namespace Support\Facades and overall small improvements
