+++ app/Modules/Users/Controllers/Admin/Roles.php
@@ -9,8 +9,6 @@
 namespace App\Modules\Users\Controllers\Admin;
 
 use Core\Config;
-use Core\View;
-use Helpers\Url;
 use Helpers\ReCaptcha;
 
 use App\Core\Controller;
@@ -24,6 +22,7 @@ use Input;
 use Redirect;
 use Session;
 use Validator;
+use View;
 
 
 class Roles extends Controller
