+++ system/Helpers/Database.php
@@ -255,7 +255,7 @@ class Database
      * @param  string  $connection
      * @return \Database\Connection
      */
-    public static function resolveConnection($connection = null)
+    public function resolveConnection($connection = null)
     {
         return static::$resolver->connection($connection);
     }
