+++ system/Helpers/Profiler.php
@@ -11,7 +11,6 @@ namespace Helpers;
 use Core\Config;
 
 use DB;
-
 use Request;
